# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Freelance Marketplace.",
        "title": "GigForge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "gigforge.jasim.online",
    "basePath": "/",
    "paths": {
...
        "/chat": {
            "get": {
                "description": "Establish a WebSocket connection for real-time chat messaging. This endpoint allows users to send and receive messages in real time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "WebSocket Chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "User_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/chat/messages/{receiver_id}": {
            "get": {
                "description": "Retrieve chat messages between the logged-in user and the specified receiver.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Get Chat Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sender User ID",
                        "name": "User_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Receiver User ID",
                        "name": "receiver_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "res.CommonRes": {
            "type": "object",
            "properties": {
                "body": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClientAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "FreelancerAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}